Added Main Menu Scene
We now have a simple but functional main menu. We can make it look
cooler later on if we have time.

For now, it's just got a simple start button that will open the outside
garden area, and an exit button that will close the application
(assuming we were running the exe, won't work in the editor).

Only one new script:
MenuScript

Don't think it breaks anything.