Added Inventory Package
I have imported the basic inventory system I was making into a separate
folder titled InventoryBase.

There are five new scripts and a bunch of other junk to worry about so i
figured it best to isolate it all where it won't break the rest of the
game.

Some of the directories referenced in my scripts need to be altered, and
I still need to integrate it into our isometric scenes. Will take care
of it Wednesday.